Centralized code, fixed bugs, deleted unnecessary lines
